refactor: UserDatabaseCredentials structure 수정
