+++ src/main/java/com/twentiethcenturygangsta/ourboard/site/OurBoardClient.java
@@ -1,7 +1,7 @@
 package com.twentiethcenturygangsta.ourboard.site;
 
 import com.twentiethcenturygangsta.ourboard.auth.UserCredentials;
-import com.twentiethcenturygangsta.ourboard.database.UserDatabaseCredentials;
+import com.twentiethcenturygangsta.ourboard.auth.UserDatabaseCredentials;
 import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
